refactored operand code
